update my fork

+++ Detectors/MUON/MCH/Simulation/include/MCHSimulation/Response.h
@@ -1,8 +1,9 @@
-// Copyright CERN and copyright holders of ALICE O2. This software is
-// distributed under the terms of the GNU General Public License v3 (GPL
-// Version 3), copied verbatim in the file "COPYING".
+// Copyright 2019-2020 CERN and copyright holders of ALICE O2.
+// See https://alice-o2.web.cern.ch/copyright for details of the copyright holders.
+// All rights not expressly granted are reserved.
 //
-// See http://alice-o2.web.cern.ch/license for full licensing information.
+// This software is distributed under the terms of the GNU General Public
+// License v3 (GPL Version 3), copied verbatim in the file "COPYING".
 //
 // In applying this license CERN does not waive the privileges and immunities
 // granted to it by virtue of its status as an Intergovernmental Organization
@@ -11,7 +12,7 @@
 #ifndef O2_MCH_SIMULATION_RESPONSE_H_
 #define O2_MCH_SIMULATION_RESPONSE_H_
 
-#include "MCHBase/Digit.h"
+#include "DataFormatsMCH/Digit.h"
 #include "MCHSimulation/Detector.h"
 #include "MCHSimulation/Hit.h"
 
@@ -38,7 +39,7 @@ class Response
   float etocharge(float edepos);
   double chargePadfraction(float xmin, float xmax, float ymin, float ymax);
   double chargefrac1d(float min, float max, double k2, double sqrtk3, double k4);
-  unsigned long response(unsigned long adc);
+  uint32_t response(uint32_t adc);
   float getAnod(float x);
   float chargeCorr();
   bool aboveThreshold(float charge) { return charge > mChargeThreshold; };
